Rename operation GBT -> GBRT everywhere
